requirements.txt
google-genai
gradio
huggingface_hub
python-dotenv